feat!: refactor project structure, rename dotflakes to devenv

+++ scripts/dotstrap/utils/dotfiles.py
@@ -5,25 +5,42 @@ import os
 import shutil
 from abc import ABC, abstractmethod
 from typing import Any
+from dataclasses import dataclass
 
 from utils.scriptargs import ScriptArgs
 
 
+class DotfileLifecycle(ABC):
+    @abstractmethod
+    def install(self):
+        pass
+
+    @abstractmethod
+    def check(self):
+        pass
+
+    @abstractmethod
+    def uninstall(self):
+        pass
+
+
+@dataclass
 class DotfilesManager:
+    dotfiles_dir: Path
+    target_dir: Path
+    dotfiles: list[DotfileLifecycle]
+
     @classmethod
     def from_script_args(cls, args: ScriptArgs):
-        return cls(args.dotfiles_directory, args.target_directory)
-
-    def __init__(self, dotfiles_directory: Path, target_directory: Path):
+        dotfiles_directory = args.dotfiles_dir
+        target_directory = args.target_dir
         config_loader = DotfileConfigLoader(dotfiles_directory, target_directory)
-
-        self.dotfiles_directory = dotfiles_directory
-        self.target_directory = target_directory
-        self.dotfiles: list[DotfileLifecycle] = [
+        dotfiles: list[DotfileLifecycle] = [
             SymlinkDotfiles(config_loader.get("dotfiles", "symlink")),
             DeprecatedDotfiles(config_loader.get("dotfiles", "deprecated")),
             ManualDotfiles(config_loader.get("dotfiles", "manual")),
         ]
+        return cls(dotfiles_directory, target_directory, dotfiles)
 
     def install_all(self):
         for dotfile in self.dotfiles:
@@ -38,10 +55,10 @@ class DotfilesManager:
             dotfile.check()
 
     def get_path(self, path: str):
-        return Path(self.dotfiles_directory, path)
+        return Path(self.dotfiles_dir, path)
 
     def get_target_path(self, path: str):
-        return Path(self.target_directory, path)
+        return Path(self.target_dir, path)
 
 
 class Dotfile:
@@ -50,7 +67,7 @@ class Dotfile:
         self.target = Path(target_directory, path)
 
     def symlink(self):
-        symlink_path(self.target, self.source)
+        symlink_path(self.source, self.target)
 
     def is_symlinked(self) -> bool:
         return self.target.is_symlink() & self.target.exists()
@@ -60,6 +77,9 @@ class Dotfile:
 
     def remove(self):
         remove_path(self.target)
+
+    def create_empty(self):
+        self.target.touch()
 
 
 class DotfileConfigLoader:
@@ -79,20 +99,6 @@ class DotfileConfigLoader:
 
     def create_dotfile(self, path: str):
         return Dotfile(path, self.dotfiles_directory, self.target_directory)
-
-
-class DotfileLifecycle(ABC):
-    @abstractmethod
-    def install(self):
-        pass
-
-    @abstractmethod
-    def check(self):
-        pass
-
-    @abstractmethod
-    def uninstall(self):
-        pass
 
 
 class SymlinkDotfiles(DotfileLifecycle):
@@ -130,8 +136,7 @@ class DeprecatedDotfiles(DotfileLifecycle):
 
     def uninstall(self):
         for dotfile in self.dotfiles:
-            if dotfile.exists():
-                dotfile.remove()
+            dotfile.remove()
 
 
 class ManualDotfiles(DotfileLifecycle):
@@ -141,7 +146,8 @@ class ManualDotfiles(DotfileLifecycle):
     def install(self):
         for dotfile in self.dotfiles:
             if not dotfile.exists():
-                print(f"Remember to create {dotfile.target}")
+                print(f"Creating {dotfile.target}")
+                dotfile.create_empty()
 
     def check(self):
         for dotfile in self.dotfiles:
@@ -154,9 +160,6 @@ class ManualDotfiles(DotfileLifecycle):
 
 
 def remove_path(path: Path):
-    if not path.exists():
-        return
-
     if path.is_symlink():
         print(f"Unlinking {path}")
         path.unlink()
@@ -167,11 +170,13 @@ def remove_path(path: Path):
         shutil.rmtree(path)
         return
 
-    print(f"Deleting file {path}")
-    os.remove(path)
+    if path.exists():
+        print(f"Deleting file {path}")
+        os.remove(path)
+        return
 
 
-def symlink_path(target: Path, source: Path):
+def symlink_path(source: Path, target: Path):
     if target.is_symlink():
         print(f"Symlink already exists: {target}")
         return
@@ -180,7 +185,7 @@ def symlink_path(target: Path, source: Path):
         print(f"Removing existing file or directory: {target}")
         remove_path(target)
 
-    print(f"Creating symlink: {target} -> {source}")
+    print(f"Creating symlink: {source} -> {target}")
     target.symlink_to(source, source.is_dir())
